refactor: original image to predict.
1. 將原圖作為預測後處理的底圖.

issue #Benson_013
