--- web_server/WSP-Pages/php/test.php
@@ -1,5 +1,0 @@
-<?php
-$test = Array();
-echo json_encode($test);
-echo gettype($test); 
-?>
