+++ web_server/WSP-Pages/wound/post_processing/edge.py
@@ -5,7 +5,7 @@ import os
 def create_edge(path):
     # superposition.png
     img_predict = cv2.imread(path + 'predict_ccl.png')
-    img_origin = cv2.imread(path + 'upload/original.png')
+    img_origin = cv2.imread(path + 'resize_original.png')
     output = cv2.addWeighted(img_origin, 0.9, img_predict, 1,0)
     cv2.imwrite(path + 'superposition.png',output)
